requirements.txt
altgraph==0.17.4
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.58.4
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
setuptools==80.9.0
six==1.17.0
